+++ aderyn_core/src/ast/impls/node/literals.rs
@@ -1,36 +1,6 @@
-use super::*;
-use crate::visitor::ast_visitor::*;
+use crate::{ast::*, visitor::ast_visitor::*};
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[serde(rename_all = "camelCase")]
-pub enum LiteralKind {
-    Bool,
-    Number,
-    String,
-    HexString,
-    Address,
-    UnicodeString, // notok
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct Literal {
-    pub hex_value: Option<String>,
-    pub value: Option<String>,
-    pub subdenomination: Option<String>,
-    pub kind: LiteralKind,
-    pub argument_types: Option<Vec<TypeDescriptions>>,
-    pub is_constant: bool,
-    pub is_l_value: bool,
-    pub is_pure: bool,
-    pub l_value_requested: bool,
-    pub type_descriptions: TypeDescriptions,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Node for Literal {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
