+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -8,3 +8,5 @@ pub mod expressions;
 pub mod functions;
 pub mod identifiers;
 pub mod import_directives;
+pub mod literals;
+pub mod modifiers;
